+++ Changelog.md
@@ -10,17 +10,26 @@ Soon....
 
 *WIP*
 
+Added caches for mod icon to load faster.
+添加了图标缓存，提升加载速度。
+
 Add Curseforge/Modrinth categories display.
 添加了Curseforge/Modrinth的类别显示。
 
 Added local mod loading progress display.
 添加了本地mod加载进度显示。
 
+Local mods without icon info will show website icons.
+没有图标的本地mod会显示存在的网络图标。
+
 Browser selector is movable toolbar now.
 浏览选择器是可移动的工具栏了。
 
 Tweak GUI of mod list.
 调整了mod列表的GUI。
+
+Mod lists now have smooth scrolling.
+mod列表现在有平滑滚动了。
 
 Resize/reposition of main window will be saved.
 主界面的调整大小和位置会被保存。